fixed command loading bug
